Fix segv in option setting
* src/confdialog.c (arg_to_str, args_are_equal): Take care of the
NO_ARG field to using the values if it has been set.
--

This is a part of the fix for bug#1413.  The other part is a fix in
gpgme.  It would be possible to do fix this only in GPA but the fix in
gpgme makes gpgme also more robust.