Change the license of some files to LGPLv2.1.
* src/filetype.c: Change license notice
* src/filetype.h: Ditto.
* src/parsetlv.c: Ditto.
* src/parsetlv.h: Ditto.
--

Those files are useful for other projects and thus we (g10 Code) allow
its use under the less restrictive LGPLv2.1.

NB: If you think that other parts of the code are also useful for
projects with different licenses requirements, please write to me.  If
g10 Code is the sole copyright holder, we may be able to change the
license for that code as well.