Fix wrong use of GPGME_EVENT_DONE.
* src/gpacontext.c (gpa_context_event_cb): Fix use of TYPE_DATA.  Add
debug output.
--

With GPGME commit c8e934b2 (2009-10-26) gpgme_io_event_done_data_t was
introduced to replace the use of gpg_error_t with GPGME_EVENT_DONE.
Unfortunately this was not documented.  Maybe at that time the event
code was considered internal and its use in GPA was not known.  Too
bad.