Add a Refresh Key function to the key manager's context menu.
* src/gpaimportbykeyidop.c: New.
* src/gpaimportbykeyidop.h: New.
* src/keymanager.c (key_manager_refresh_keys): New.
(key_manager_action_new): Add ServerRefresh Action top the popup.
--

This is currently limited to one key.  It won't be complicated to
allow several keys but this has not been done because it is better
to first write code to get an array of gpgme key objects from the key
manager's selection instead of passing the selection (which has only
opaque types) to the import functions.