Add very basic support for PIV cards.
* src/cm-piv.c: New.
* src/cm-piv.h: New.
* src/Makefile.am (gpa_cardman_sources): Add them.
* src/cardman.c: Support PIV.
* src/keytable.c (gpa_keytable_ensure): New.
(gpa_keytable_lookup_key): Replace code by a call the new func.
--

Take care: I wrote that code years ago and it had not been updated to
newer GTK+ versions, debugged, or even has a proper layout.  But it is
better to have it in the repo than only in my stash.